refactor: price의 타입을 long에서 int로 변경

+++ gc-coffee/src/main/java/com/example/gccoffee/repository/OrderJdbcRepository.java
@@ -63,7 +63,7 @@ public class OrderJdbcRepository implements OrderRepository {
                   UUID.fromString(resultSet.getString("productId")),
                   resultSet.getString("name"),
                   Category.valueOf(resultSet.getString("category")),
-                  resultSet.getLong("price"),
+                  resultSet.getInt("price"),
                   resultSet.getString("description")
           );
